+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Version 4.0.0
 -------------
-_23 Jul 2019_
+_24 Jul 2019_
 
 MockUIAlertController is dead. Long live ViewControllerPresentationSpy!